+++ tests/bundles/graphene/test_mutations.py
@@ -4,8 +4,8 @@ from flask_unchained import unchained
 
 from ._bundles.graphene_bundle.graphql import types
 
-parent_manager = unchained.get_service_local_proxy('parent_manager')
-child_manager = unchained.get_service_local_proxy('child_manager')
+parent_manager = unchained.get_local_proxy('parent_manager')
+child_manager = unchained.get_local_proxy('child_manager')
 
 
 CREATE_PARENT = '''