+++ tests/bundles/graphene/test_schema.py
@@ -3,8 +3,8 @@ import pytest
 from flask_unchained import unchained
 
 
-parent_manager = unchained.get_service_local_proxy('parent_manager')
-child_manager = unchained.get_service_local_proxy('child_manager')
+parent_manager = unchained.get_local_proxy('parent_manager')
+child_manager = unchained.get_local_proxy('child_manager')
 
 
 GET_CHILD = '''